Check object map section checksum
Before the change object map section checksum was ignored. As a result,
if we started to parse object map sections past its actual contents, we could
read some garbage.

Given that we the code computes checksum correctly, we now use it to verify
if we are reading the right data. When we find invalid checksum, then we treat
it as file error.